Move Trace classes to separate module
